+++ gen/com/intellij/idea/plugin/hybris/flexibleSearch/psi/FlexibleSearchDefinedTableName.java
@@ -30,4 +30,7 @@ public interface FlexibleSearchDefinedTableName extends PsiElement {
   @NotNull
   String getTableName();
 
+  @Nullable
+  FlexibleSearchTableAliasName getTableAlias();
+
 }