+++ gen/com/intellij/idea/plugin/hybris/flexibleSearch/psi/FlexibleSearchYColumnName.java
@@ -24,13 +24,15 @@ package com.intellij.idea.plugin.hybris.flexibleSearch.psi;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
+import java.util.Collection;
+import kotlin.Pair;
 
 public interface FlexibleSearchYColumnName extends PsiElement {
 
   @Nullable
-  FlexibleSearchDefinedTableName getTable();
+  Pair<FlexibleSearchDefinedTableName, FlexibleSearchTableAliasName> getTableToAlias();
 
   @NotNull
-  List<FlexibleSearchTableAliasName> getTableAliases();
+  Collection<FlexibleSearchTableAliasName> getTableAliases();
 
 }