+++ src/com/intellij/idea/plugin/hybris/flexibleSearch/settings/FlexibleSearchSettings.kt
@@ -25,6 +25,7 @@ import com.intellij.openapi.components.BaseState
 data class FlexibleSearchSettings(
     var verifyCaseForReservedWords: Boolean = true,
     var verifyUsedTableAliasSeparator: Boolean = true,
+    var fallbackToTableNameIfNoAliasProvided: Boolean = true,
     var defaultCaseForReservedWords: ReservedWordsCase = ReservedWordsCase.UPPERCASE,
 
     var completion: FlexibleSearchCompletionSettings = FlexibleSearchCompletionSettings(),